requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas
numpy
scikit-learn
fastapi
uvicorn
python-jose
sqlalchemy
passlib
bcrypt
transformers
deep-translator
